data_lake/requirements.txt
django
djangorestframework
pyspark
python-dotenv